+++ nym-vpn-core/crates/nym-vpn-api-client/src/client.rs
@@ -10,7 +10,7 @@ use serde::{de::DeserializeOwned, Serialize};
 use crate::{
     error::{Result, VpnApiClientError},
     request::{CreateSubscriptionRequestBody, RegisterDeviceRequestBody, RequestZkNymRequestBody},
-    responses::{
+    response::{
         NymDirectoryGatewayCountriesResponse, NymDirectoryGatewaysResponse, NymVpnAccountResponse,
         NymVpnAccountSummaryResponse, NymVpnDevice, NymVpnDevicesResponse, NymVpnSubscription,
         NymVpnSubscriptionResponse, NymVpnZkNym, NymVpnZkNymResponse,
