+++ nym-vpn-core/crates/nym-vpn-api-client/src/jwt.rs
@@ -1,3 +1,6 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
 use std::fmt;
 
 use nym_crypto::asymmetric::ed25519::KeyPair;
