+++ nym-vpn-core/crates/nym-vpn-api-client/src/lib.rs
@@ -1,7 +1,7 @@
 // Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-pub mod responses;
+pub mod response;
 pub mod types;
 
 pub(crate) mod jwt;
